keep a test script in base directory
